requirements.txt
pymupdf4llm==0.0.24
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
autogen-ext[ollama]==0.5.7
accelerate==1.6.0
markdown==3.7
# pandas==2.2.0
# transformers==4.52.3
# gradio==5.31.0
langchain-ollama==0.3.6
ollama==0.5.1
langchain-community==0.3.27
pdfplumber==0.11.7
python-multipart